Use a nonbreaking space before year.
In getDepartmentMembers(), Now return a (utf8) nonbreaking space
between every name and its year (if there is one), instead of a
regular space. This is intended to make the mast fit better and avoid
pargraph orphans that consist solely of years.

For instance, we almost had:

  PRODUCTION STAFF
  ----------------
  Editors: Annia Pan ’15, Anthony Yu ’16; Staff:
  Judy Hsiang ’12, Fareeha Safir ’13, Kezi Cheng
  ’15, Esme Rhine ’15, Will Conway ’16, Kath Xu
  ’16.

With nonbreaking spaces, we get:

  PRODUCTION STAFF
  ----------------
  Editors: Annia Pan ’15, Anthony Yu ’16;
  Staff: Judy Hsiang ’12, Fareeha Safir ’13, Kezi
  Cheng ’15, Esme Rhine ’15, Will Conway ’16,
  Kath Xu ’16.

I was initially concerned that this might cause copyfitting problems
(bigger word blocks) and had intended to only apply the nonbreaking
space to the space before the final year of a section. But tests show
that does not seem to be a problem, so its simpler to go with all
spaces before years.

One could also go a step further and do this for all spaces internal
to a name. Expiriments shwo this is a bad idea: copyfitting suffers,
and there are significant rivers. Years stack up at the right margin
and look awkward. And the mast flows to more lines. All of this is because
we're not hyphenating but we are justifying, a tricky thing. Fortunately
it works for us.

An alternative heuristic that looks at the length of words/names might
be appropriate, e.g. "Kath<NONBREAKING>Xu" but "Leon<SPACE>Lin", if we
wanted to pursue this. I think it's a non-problem right now, and certainly
better than the prior state.

Note that this change affects both taggedTextMast.php and htmlMast.php;
it is really intended for the InDesign (tagged text) mast, but I don't
think modern any browsers fail to deal with utf-8 nonbreaking spaces,
so it's fine to do it globally.

This copyfitting in the masthead has been bothering me for years. Glad
to finally get it off my chest.


git-svn-id: file:///srv/svn/trunk/3m@2224 8d44d787-80f3-42f6-8d0b-07ebc01cdc8b